# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Cms.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:61940/"
    }
  ],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Establish an authenticated session.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfUserView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfUserView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfUserView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Change password and invalidate the current session.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfboolean"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfboolean"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfboolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/themes/preview": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Validate a non-persistent appearance preview.",
        "parameters": [
          {
            "name": "themeId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accentColor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "heroTitle",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "heroDescription",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfThemeView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfThemeView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfThemeView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/themes/active": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Atomically save and activate a packaged theme.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeApplyInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeApplyInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeApplyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfThemesView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfThemesView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfThemesView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/contents": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List content drafts.",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "post"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "recent"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfPageResultOfContentView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfPageResultOfContentView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfPageResultOfContentView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create a draft.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ContentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfContentView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfContentView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfContentView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/templates/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Read a published reusable layout through an authenticated editorial session.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfContentView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfContentView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfContentView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/blocks/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Read a published reusable block for the page builder.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfContentView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfContentView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfContentView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/blocks/{id}/references": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List retained uses of a synchronized block.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfIReadOnlyListOfBlockReference"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfIReadOnlyListOfBlockReference"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfIReadOnlyListOfBlockReference"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseOfUserView": {
        "required": [
          "code",
          "message",
          "data",
          "traceId"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserView"
              }
            ]
          },
          "traceId": {
            "type": "string"
          }
        },
        "description": "Consistent HTTP response envelope."
      },
...
      "LoginInput": {
        "required": [
          "username",
          "password",
          "captchaId",
          "captchaCode"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "captchaId": {
            "type": "string"
          },
          "captchaCode": {
            "type": "string"
          }
        },
        "description": "Login credentials."
      },
...
    },
    "securitySchemes": {
      "IntegrationToken": {
        "type": "http",
        "description": "Expiring CMS access token issued by an administrator. HTTPS required; no cookie or CSRF token is used on integration endpoints.",
        "scheme": "bearer"
      }
    }
  },
...
}